backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
python-multipart>=0.0.9
python-dotenv==1.0.0
aiosqlite==0.21.0
reportlab==4.4.5